requirements.txt
discord.py==1.7.3
Flask==3.0.3
